+++ btcstaking-tracker/stakingeventwatcher/stakingeventwatcher.go
@@ -188,6 +188,7 @@ func (sew *StakingEventWatcher) handleNewBlocks(blockNotifier *notifier.BlockEpo
 				panic(fmt.Errorf("received negative block height: %d", block.Height))
 			}
 			sew.currentBestBlockHeight.Store(uint32(block.Height))
+			sew.metrics.BestBlockHeight.Set(float64(block.Height))
 			sew.logger.Debugf("Received new best btc block: %d", block.Height)
 
 			if err := sew.checkSpend(); err != nil {
@@ -326,10 +327,12 @@ func (sew *StakingEventWatcher) syncedWithBabylon() (bool, error) {
 		return false, err
 	}
 
-	currentBtcNodeHeight := sew.currentBestBlockHeight.Load()
+	sew.metrics.BtcLightClientHeight.Set(float64(btcLightClientTipHeight))
 
+	currentBtcNodeHeight := sew.currentBestBlockHeight.Load()
 	if currentBtcNodeHeight < btcLightClientTipHeight {
-		sew.logger.Debugf("btc light client tip height is %d, connected node best block height is %d. Waiting for node to catch up", btcLightClientTipHeight, currentBtcNodeHeight)
+		sew.logger.Debugf("btc light client tip height is %d, connected node best block height is %d. Waiting for node to catch up",
+			btcLightClientTipHeight, currentBtcNodeHeight)
 
 		return false, nil
 	}
@@ -656,7 +659,7 @@ func (sew *StakingEventWatcher) handlerVerifiedDelegations() {
 func (sew *StakingEventWatcher) checkBtcForStakingTx() {
 	params, err := sew.babylonNodeAdapter.Params()
 	if err != nil {
-		sew.logger.Errorf("error getting tx params %v", err)
+		sew.logger.Errorf("error getting tx params: %s", err.Error())
 
 		return
 	}
@@ -669,7 +672,7 @@ func (sew *StakingEventWatcher) checkBtcForStakingTx() {
 
 		details, status, err := sew.btcClient.TxDetails(&txHash, del.StakingTx.TxOut[del.StakingOutputIdx].PkScript)
 		if err != nil {
-			sew.logger.Debugf("error getting tx: %v, err: %v", txHash, err)
+			sew.logger.Errorf("error getting tx: %v, err: %v", txHash, err)
 
 			continue
 		}
@@ -692,7 +695,7 @@ func (sew *StakingEventWatcher) checkBtcForStakingTx() {
 
 		proof, err := ib.GenSPVProof(int(details.TxIndex))
 		if err != nil {
-			sew.logger.Debugf("error making spv proof %s", err)
+			sew.logger.Warnf("error making spv proof %s", err)
 
 			continue
 		}
@@ -704,7 +707,7 @@ func (sew *StakingEventWatcher) checkBtcForStakingTx() {
 		}
 
 		if err := sew.pendingTracker.UpdateActivation(txHash, true); err != nil {
-			sew.logger.Debugf("error updating activation in pending tracker tx: %v", txHash)
+			sew.logger.Warnf("error updating activation in pending tracker tx: %v", txHash)
 			sew.activationLimiter.Release(1) // in probable edge case, insure we release the sem
 
 			continue
@@ -733,12 +736,12 @@ func (sew *StakingEventWatcher) activateBtcDelegation(
 	defer sew.latency("activateBtcDelegation")()
 	defer func() {
 		if err := sew.pendingTracker.UpdateActivation(stakingTxHash, false); err != nil {
-			sew.logger.Debugf("err updating activation in pending tracker tx: %v", stakingTxHash)
+			sew.logger.Warnf("err updating activation in pending tracker tx: %v", stakingTxHash)
 		}
 	}()
 
-	if err := sew.waitForRequiredDepth(ctx, stakingTxHash, &inclusionBlockHash, requiredDepth); err != nil {
-		sew.logger.Warnf("exceeded waiting for required depth for tx: %s, will try later: err %v", stakingTxHash.String(), err)
+	if err := sew.waitForRequiredDepth(ctx, &inclusionBlockHash, requiredDepth); err != nil {
+		sew.logger.Warnf("exceeded waiting for required depth for tx: %s, will try later. Err: %v", stakingTxHash.String(), err)
 
 		if err := sew.verifiedInsufficientConfTracker.AddEmptyDelegation(stakingTxHash); err == nil {
 			sew.metrics.NumberOfVerifiedInsufficientConfDelegations.Set(float64(sew.verifiedInsufficientConfTracker.Count()))
@@ -787,7 +790,7 @@ func (sew *StakingEventWatcher) activateBtcDelegation(
 		sew.pendingTracker.RemoveDelegation(stakingTxHash)
 		sew.metrics.NumberOfVerifiedDelegations.Dec()
 
-		sew.logger.Debugf("staking tx activated %s", stakingTxHash.String())
+		sew.logger.Infof("staking tx activated %s", stakingTxHash.String())
 
 		if _, exists := sew.verifiedSufficientConfTracker.GetDelegation(stakingTxHash); exists {
 			sew.verifiedSufficientConfTracker.RemoveDelegation(stakingTxHash)
@@ -809,7 +812,6 @@ func (sew *StakingEventWatcher) activateBtcDelegation(
 
 func (sew *StakingEventWatcher) waitForRequiredDepth(
 	ctx context.Context,
-	stakingTxHash chainhash.Hash,
 	inclusionBlockHash *chainhash.Hash,
 	requiredDepth uint32,
 ) error {
@@ -843,10 +845,6 @@ func (sew *StakingEventWatcher) waitForRequiredDepth(
 		retry.MaxDelay(sew.cfg.RetrySubmitUnbondingTxInterval),
 		retry.MaxJitter(sew.cfg.RetryJitter),
 		retry.LastErrorOnly(true), // let's avoid high log spam
-		retry.OnRetry(func(n uint, err error) {
-			sew.logger.Debugf("waiting for staking tx: %s to be k-deep. Current[%d], required[%d]. "+
-				"Attempt: %d. Err: %v", stakingTxHash, depth, requiredDepth, n, err)
-		}),
 	)
 }
 
@@ -855,7 +853,6 @@ func (sew *StakingEventWatcher) latency(method string) func() {
 
 	return func() {
 		duration := time.Since(startTime)
-		sew.logger.Debugf("execution time for method: %s, duration: %s", method, duration.String())
 		sew.metrics.MethodExecutionLatency.WithLabelValues(method).Observe(duration.Seconds())
 	}
 }